Added Tunnels
Had to change from using canMoveTo to whereToMove.
Because of that I made new tests for obth Tunnels and Globe.

I had to somehow return grass and tunnels from objectAt so I made WorldElement interface and 2 dumb classes that implement it: Plant and Tunnel.

Maybe it is better to make an abstract class?
The problem was caused by MapVisualizer because I am not sure how we will use it.